pyproject.toml
[tool.poetry]
name = "magic-deck-set-tool"
version = "0.1.0"
description = ""
authors = ["Ramonnn <29675067+Ramonnn@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
